+++ src/consumer/component/consumer-component.js
@@ -7,7 +7,7 @@ import { pathNormalizeToLinux } from '../../utils';
 import createSymlinkOrCopy from '../../utils/fs/create-symlink-or-copy';
 import ComponentBitConfig from '../bit-json';
 import { Dist, License, SourceFile } from '../component/sources';
-import type ConsumerBitConfig from '../bit-json/consumer-bit-config';
+import type ConsumerBitConfig from '../bit-config/consumer-bit-config';
 import type Consumer from '../consumer';
 import BitId from '../../bit-id/bit-id';
 import type Scope from '../../scope/scope';
@@ -50,7 +50,7 @@ import type { RawTestsResults } from '../specs-results/specs-results';
 import { paintSpecsResults } from '../../cli/chalk-box';
 import ExternalTestErrors from './exceptions/external-test-errors';
 import GeneralError from '../../error/general-error';
-import AbstractBitConfig from '../bit-json/abstract-bit-config';
+import AbstractBitConfig from '../bit-config/abstract-bit-config';
 import { Analytics } from '../../analytics/analytics';
 import type { PackageJsonInstance } from './package-json';
 import { componentIssuesLabels } from '../../cli/templates/component-issues-template';
