+++ src/extensions/env-extension.js
@@ -13,7 +13,7 @@ import { Repository } from '../scope/objects';
 import { pathJoinLinux } from '../utils';
 import removeFilesAndEmptyDirsRecursively from '../utils/fs/remove-files-and-empty-dirs-recursively';
 import type { PathOsBased } from '../utils/path';
-import type { EnvExtensionObject } from '../consumer/bit-json/abstract-bit-config';
+import type { EnvExtensionObject } from '../consumer/bit-config/abstract-bit-config';
 import { ComponentWithDependencies } from '../scope';
 import { Analytics } from '../analytics/analytics';
 import ExtensionGetDynamicPackagesError from './exceptions/extension-get-dynamic-packages-error';
@@ -22,7 +22,7 @@ import type TesterExtension from './tester-extension';
 import { COMPONENT_ORIGINS, COMPILER_ENV_TYPE, TESTER_ENV_TYPE } from '../constants';
 import type { ComponentOrigin } from '../consumer/bit-map/component-map';
 import type ConsumerComponent from '../consumer/component';
-import type ConsumerBitConfig from '../consumer/bit-json/consumer-bit-config';
+import type ConsumerBitConfig from '../consumer/bit-config/consumer-bit-config';
 import type ComponentBitConfig from '../consumer/bit-json';
 import logger from '../logger/logger';
 import { Dependencies } from '../consumer/component/dependencies';
